# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Calls",
        "title": "Desktop Eamil",
        "termsOfService": "http://fromkeith.com",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5173",
    "basePath": "/api",
    "paths": {
...
        "/messages/push": {
            "post": {
                "description": "Sync endpoint to push client changes to messages for this account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Update Messages",
                "parameters": [
                    {
                        "description": "Push Message Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PushMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PushMessageResponse"
                        }
                    }
                }
            }
        },
...
        "/threads/pull": {
            "get": {
                "description": "Sync endpoint to pull all changes to threads for this account.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Get Threads",
                "parameters": [
                    {
                        "type": "string",
                        "description": "threadId",
                        "name": "threadId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Last updated time",
                        "name": "updatedAt",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Batch size",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PullThreadResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "CheckpointThreads": {
            "type": "object",
            "required": [
                "threadId",
                "updatedAt"
            ],
            "properties": {
                "threadId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "GmailEntry": {
            "type": "object",
            "required": [
                "additionalReceivers",
                "categories",
                "createdAt",
                "headers",
                "historyId",
                "internalDate",
                "isDeleted",
                "labels",
                "messageId",
                "receivedAt",
                "receiver",
                "revisionCount",
                "sender",
                "snippet",
                "subject",
                "tags",
                "threadId",
                "todos",
                "updatedAt",
                "userId"
            ],
            "properties": {
                "additionalReceivers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/PersonInfo"
                        }
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "historyId": {
                    "type": "integer"
                },
                "internalDate": {
                    "description": "The internal message creation timestamp (epoch ms), which determines ordering in the inbox. For normal SMTP-received email, this represents the time the message was originally accepted by Google, which is more reliable than the Date header. However, for API-migrated mail, it can be configured by client to be based on the Date header.",
                    "type": "integer"
                },
                "isDeleted": {
                    "description": "generated by us",
                    "type": "boolean"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "messageId": {
                    "type": "string"
                },
                "receivedAt": {
                    "description": "probaly use internalDate",
                    "type": "string"
                },
                "receiver": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PersonInfo"
                    }
                },
                "replyTo": {
                    "$ref": "#/definitions/PersonInfo"
                },
                "revisionCount": {
                    "type": "integer"
                },
                "sender": {
                    "$ref": "#/definitions/PersonInfo"
                },
                "snippet": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "threadId": {
                    "type": "string"
                },
                "todos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "description": "For Sync + Conflict Resolution",
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "MessageBasic": {
            "type": "object",
            "required": [
                "internalDate",
                "labels",
                "messageId"
            ],
            "properties": {
                "internalDate": {
                    "type": "integer"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "messageId": {
                    "type": "string"
                },
                "sender": {
                    "$ref": "#/definitions/PersonInfo"
                },
                "snippet": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "PersonInfo": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "PullThreadResponse": {
            "type": "object",
            "required": [
                "checkpoint",
                "threads"
            ],
            "properties": {
                "checkpoint": {
                    "$ref": "#/definitions/CheckpointThreads"
                },
                "threads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Thread"
                    }
                }
            }
        },
        "PushMessageRequest": {
            "type": "object",
            "required": [
                "rows"
            ],
            "properties": {
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PushMessageRow"
                    }
                }
            }
        },
        "PushMessageResponse": {
            "type": "object",
            "required": [
                "conflicts"
            ],
            "properties": {
                "conflicts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/GmailEntry"
                    }
                }
            }
        },
        "PushMessageRow": {
            "type": "object",
            "properties": {
                "assumedMasterState": {
                    "$ref": "#/definitions/GmailEntry"
                },
                "newDocumentState": {
                    "$ref": "#/definitions/GmailEntry"
                }
            }
        },
...
        "Thread": {
            "type": "object",
            "required": [
                "categories",
                "messages",
                "mostRecentInternalDate",
                "tags",
                "threadId",
                "updatedAt"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MessageBasic"
                    }
                },
                "mostRecentInternalDate": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "threadId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}